module-3/sagemaker-projects-portfolio/service_catalog/sm_projects_products/building/build_model_bank_marketing/seed_code/scripts/requirements.txt
boto3==1.28.57
botocore>=1.31.57
sagemaker<=2.187.0
numpy<=1.26.3
pandas==1.1.3
awswrangler<=3.4.2
pyyaml==6.0.1
python-dateutil<=2.8.1
